requirements-dev.txt
Django==1.8.6
Jinja2==2.8
Markdown==2.6.4
MarkupSafe==0.23
PyJWT==1.4.0
PyYAML==3.11
Sphinx==1.3.1
argparse==1.4.0
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
certifi==2015.9.6.2
cffi==1.3.0
coverage==4.0.2
cryptography==1.1
decorator==4.0.4
django-annoying==0.8.4
django-braces==1.8.1
django-cors-headers==1.1.0
django-countries==3.4.1
django-debug-toolbar==1.4
django-exclusivebooleanfield==0.3.0
django-extensions==1.5.9
django-filter==0.11.0
-e git+git@github.com:abdulhaq-e/django-linguo.git@e493d0d67806d8b3dbc4d222d2116791cb54c37f#egg=django_linguo-get_language_fix_django_1.8
django-oauth-toolkit==0.9.0
django-rest-swagger==0.3.4
-e git+https://github.com/llazzaro/django-scheduler.git@e82ee1ffdee45a4dd9113ae23ca52c2bd6ffa809#egg=django_scheduler-master
djangorestframework==3.3.1
-e git+git@github.com:abdulhaq-e/django-rest-framework-json-api.git@ecef2cdcd945855278f44b7da98cb9b499e23e50#egg=djangorestframework_jsonapi-my_updated_branch
-e git+git@github.com:abdulhaq-e/django-rest-framework-jwt.git@f07772769c0f72f971daefe1d9426a90ea099a63#egg=djangorestframework_jwt-upstream/HEAD
enum34==1.0.4
factory-boy==2.6.0
fake-factory==0.5.3
functools32==3.2.3.post2
future==0.15.2
icalendar==3.9.1
idna==2.0
importmagic==0.1.3
inflection==0.3.1
ipaddress==1.0.14
ipython==4.0.0
ipython-genutils==0.1.0
jedi==0.9.0
jsonschema==2.5.1
mod-wsgi==4.4.21
oauthlib==1.0.3
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
py==1.4.30
pyasn1==0.1.9
pycparser==2.14
pytest==2.8.2
pytest-factoryboy==1.1.4
python-dateutil==2.4.2
python-openid==2.2.5
pytz==2015.7
pyzmq==15.0.0
requests-oauthlib==0.5.0
simplegeneric==0.8.1
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.1.18
terminado==0.5
tornado==4.3
traitlets==4.0.0
wsgiref==0.1.2
